# swagger.yaml
openapi: 3.0.0
info:
  title: Training
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:3000
  variables: {}
paths:
  /auth/updateUser:
    patch:
      tags:
      - Auth
      summary: Update User
      operationId: UpdateUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateUserRequest'
              - example:
                  name: test user
            example:
              name: test user
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /auth/getAll:
    get:
      tags:
      - Auth
      summary: Get All USers
      operationId: GetAllUSers
      parameters:
      - name: limit
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 10
      - name: offset
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 0
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /admin/getAllUser:
    get:
      tags:
      - Admin
      summary: Get All Users
      operationId: GetAllUsers
      parameters:
      - name: page
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 1
      - name: limit
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 20
      - name: searchAttribute
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: name
      - name: searchValue
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: test
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /admin/removeUser/103e8e89-7fa4-485b-a107-6a1869f904d2:
    patch:
      tags:
      - Admin
      summary: Remove User
      operationId: RemoveUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
      - name
      type: object
      properties:
        name:
          type: string
      example:
        name: test user
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...